Move code to src/conversions/indexmap.rs
