+++ src/conversions/mod.rs
@@ -1,5 +1,6 @@
 //! This module contains conversions between various Rust object and their representation in Python.
 
 mod array;
+mod indexmap;
 mod osstr;
 mod path;
